authorization commit - checking user ID
added checking user ID in endpoints
added UserContextService - allows to access the user credentials

+++ Services/WatchedAnnoucementService.cs
@@ -18,10 +18,12 @@ namespace imotoAPI.Services
     public class WatchedAnnoucementService : IWatchedAnnoucementService
     {
         private readonly ImotoDbContext _dbContext;
+        private readonly IUserContextService _userContextService;
 
-        public WatchedAnnoucementService(ImotoDbContext dbContext)
+        public WatchedAnnoucementService(ImotoDbContext dbContext, IUserContextService userContextService)
         {
             _dbContext = dbContext;
+            _userContextService = userContextService;
         }
 
         public WatchedAnnoucementReturnDto Add(WatchedAnnoucementDto dto)
@@ -55,6 +57,9 @@ namespace imotoAPI.Services
 
             if (watchedAnnoucement is null)
                 throw new NotFoundException("Not found");
+
+            if (_userContextService.GetUserRole == "użytkownik" && _userContextService.GetUserId != watchedAnnoucement.UserId)
+                throw new ForbidException("");
 
             _dbContext.WatchedAnnoucements.Remove(watchedAnnoucement);
             _dbContext.SaveChanges();
